Magic strings in API + refactor of dashboard menu
