+++ src/components/dashboard/Dashboard.jsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { useState } from "react";
 import { useUserContext } from "../../store/UserProvider";
-import Profile from "./Profile/Profile";
+import Profile from "./profile/Profile";
 import { DASHBOARD_SCREENS } from "../../constants";
 import DashboardMenu from "./DashboardMenu";
-import UserImport from "./UserImport";
-import Welcome from "./Welcome";
-import RentalHistory from "./RentalHistory";
+import UserImport from "./importUsers/UserImport";
+import Welcome from "./welcome/Welcome";
+import RentalHistory from "./userRentalHistory/RentalHistory";
 
 function Dashboard({ noUserDetected }) {
 
@@ -29,9 +29,9 @@ function Dashboard({ noUserDetected }) {
       {currentDashboardScreen === DASHBOARD_SCREENS.IMPORT_USERS && (
         <UserImport/>
       )}
-        {currentDashboardScreen === DASHBOARD_SCREENS.RENTAL_HISTORY && (
-          <RentalHistory/>
-        )}
+      {currentDashboardScreen === DASHBOARD_SCREENS.RENTAL_HISTORY && (
+        <RentalHistory/>
+      )}
     </main>
   );
 }
