+++ src/components/dashboard/profile/ProfileForm.jsx
@@ -8,6 +8,7 @@ import { updateUserData } from "../../../services/api";
 import { ProfileInput } from "./ProfileInput";
 import Modal from "../../utils/Modal";
 import { AvatarInput } from "./AvatarInput";
+import { ERROR_TYPES, RESPONSE_TYPES } from "../../../constants";
 
 export function ProfileForm({ user, onUserUpdate }) {
   const [errors, setErrors] = useState({});
@@ -17,58 +18,55 @@ export function ProfileForm({ user, onUserUpdate }) {
 
   async function handleSubmit(e) {
     e.preventDefault();
-
+  
     setIsLoading(true);
     setIsModalOpen(true);
     setModalMessage("Actualitzant dades...");
     setErrors({});
-
+  
     const formData = new FormData(e.target);
     const file = formData.get("avatar");
     const email = formData.get("email");
     const telefon = formData.get("telefon");
-
+  
     const newErrors = {};
-
+  
     if (!isValidEmail(email)) {
       newErrors.email = "Email no té un format válid.";
     }
-
+  
     if (!isValidTelefon(telefon)) {
       newErrors.telefon = "El teléfon ha de tenir 9 dígits.";
     }
-
+  
     if (file && file.name !== "" && !isValidFileType(file)) {
-      newErrors.avatar =
-        "Format de imatge invàlid. Només poden ser JPG, JPEG i PNG.";
+      newErrors.avatar = "Format de imatge invàlid. Només poden ser JPG, JPEG i PNG.";
     }
-
+  
     if (Object.keys(newErrors).length > 0) {
       setErrors(newErrors);
-
       setModalMessage("Hi ha errors");
       setIsLoading(false);
-
       return;
     }
-
+  
     const token = localStorage.getItem("authToken");
     const res = await updateUserData(formData, token);
-
+  
     if (res.success) {
-      if (res.data.type === "success_modify") {
+      if (res.data.type === RESPONSE_TYPES.SUCCESS_MODIFY) {
         onUserUpdate(res.data.userData);
         setModalMessage("Les dades s'han actualitzat correctament");
       } else {
         setModalMessage("No hi han hagut canvis");
       }
     } else {
-      if (res.type === "error_at_data") {
+      if (res.type === ERROR_TYPES.ERROR_AT_DATA) {
         setErrors(res.error.formErrors);
       }
       setModalMessage("S'han produit errors");
     }
-
+  
     setIsLoading(false);
   }
